pyproject.toml
[tool.poetry]
name = "sql-testing-demo"
version = "0.1.0"
description = "Demo project for unified SQL testing across DuckDB and Snowflake with SQL DDL schema management"
authors = ["Vyncent Yaltar <vyncent.yaltar@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.0"
sqlglot = "^20.0.0"
pytest = "^7.0.0"
duckdb = "^0.9.0"
pandas = "^2.0.0"
snowflake-sqlalchemy = { version = "^1.5.0", optional = true }
python-dotenv = "^1.0.0"
sqlacodegen = "^3.0.0"
duckdb-engine = "^0.17.0"

[tool.poetry.extras]
snowflake = ["snowflake-sqlalchemy"]
codegen = ["sqlacodegen"]
full = ["snowflake-sqlalchemy", "sqlacodegen"]

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    # Ignore deprecation warnings from Snowflake connector dependencies
    "ignore::DeprecationWarning:snowflake.connector.vendored.urllib3.contrib.pyopenssl",
    # Ignore other common Snowflake-related warnings
    "ignore::DeprecationWarning:snowflake.connector.*",
    # Keep other warnings visible
    "default::UserWarning",
    "default::FutureWarning"
]

[tool.black]
line-length = 88
target-version = ['py310']